requirements.txt
tqdm
requests
schedule
apscheduler
python-dotenv
python-telegram==0.18.0